Move index.js to known location, include lint auto-fix changes
